requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
ecdsa==0.19.1
frozenlist==1.7.0
geoip2==5.1.0
git-filter-repo==2.47.0
gunicorn==23.0.0
idna==3.10
Markdown==3.7
MarkupSafe==3.0.2
maxminddb==2.7.0
multidict==6.6.2
packaging==24.2
phonenumbers==9.0.8
pillow==11.1.0
propcache==0.3.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
python-http-client==3.3.7
requests==2.32.4
sendgrid==6.12.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
